+++ Modules/SharePointDsc/DSCResources/MSFT_SPManagedMetadataServiceApp/MSFT_SPManagedMetaDataServiceApp.psm1
@@ -56,8 +56,8 @@ function Get-TargetResource
         { 
             return $nullReturn 
         }
-        $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "Managed Metadata Service" 
+        $serviceApp = $serviceApps | Where-Object -FilterScript {
+            $_.GetType().FullName -eq "Microsoft.SharePoint.Taxonomy.MetadataWebServiceApplication"
         }
 
         if ($null -eq $serviceApp)
@@ -189,8 +189,8 @@ function Set-TargetResource
                 $params = $args[0]
                 
                 $serviceApp = Get-SPServiceApplication -Name $params.Name `
-                    | Where-Object -FilterScript { 
-                        $_.TypeName -eq "Managed Metadata Service" 
+                    | Where-Object -FilterScript {
+                        $_.GetType().FullName -eq "Microsoft.SharePoint.Taxonomy.MetadataWebServiceApplication" 
                 }
                 $appPool = Get-SPServiceApplicationPool -Identity $params.ApplicationPool
                 Set-SPMetadataServiceApplication -Identity $serviceApp -ApplicationPool $appPool
@@ -208,7 +208,7 @@ function Set-TargetResource
             $params = $args[0]
             
             $serviceApp = Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript {
-                $_.TypeName -eq "Managed Metadata Service"  
+                $_.GetType().FullName -eq "Microsoft.SharePoint.Taxonomy.MetadataWebServiceApplication"  
             }
             Remove-SPServiceApplication $serviceApp -Confirm:$false
         }