+++ Tests/Unit/SharePointDsc/SharePointDsc.SPBCSServiceApp.Tests.ps1
@@ -21,8 +21,9 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
             DatabaseServer = "TestServer\Instance"
             Ensure = "Present"
         }
-        Import-Module (Join-Path ((Resolve-Path $PSScriptRoot\..\..\..).Path) "Modules\SharePointDsc")
+        $getTypeFullName = "Microsoft.SharePoint.BusinessData.SharedService.BdcServiceApplication"
 
+        Import-Module (Join-Path ((Resolve-Path $PSScriptRoot\..\..\..).Path) "Modules\SharePointDsc")
         
         Mock Invoke-SPDSCCommand { 
             return Invoke-Command -ScriptBlock $ScriptBlock -ArgumentList $Arguments -NoNewScope
@@ -34,7 +35,6 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
         Mock Remove-SPServiceApplication { }
         
         Context "When no service applications exist in the current farm and it should" {
-
             Mock Get-SPServiceApplication { return $null }
             Mock New-SPBusinessDataCatalogServiceApplication { }
 
@@ -54,10 +54,15 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
         }
 
         Context "When service applications exist in the current farm with the same name but is the wrong type" {
-
-            Mock Get-SPServiceApplication { return @(@{
-                TypeName = "Some other service app type"
-            }) }
+            Mock Get-SPServiceApplication {
+                $spServiceApp = [pscustomobject]@{
+                    DisplayName = $testParams.Name
+                }
+                $spServiceApp | Add-Member ScriptMethod GetType { 
+                    return @{ FullName = "Microsoft.Office.UnKnownWebServiceApplication" } 
+                } -PassThru -Force
+                return $spServiceApp
+            }
 
             It "returns absent from the Get method" {
                 (Get-TargetResource @testParams).Ensure | Should Be "Absent"
@@ -67,16 +72,19 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
         }
 
         Context "When a service application exists and it should, and is also configured correctly" {
-            Mock Get-SPServiceApplication { 
-                return @(@{
-                    TypeName = "Business Data Connectivity Service Application"
+            Mock Get-SPServiceApplication {
+                $spServiceApp = [pscustomobject]@{
                     DisplayName = $testParams.Name
                     ApplicationPool = @{ Name = $testParams.ApplicationPool }
                     Database = @{
                         Name = $testParams.DatabaseName
                         Server = @{ Name = $testParams.DatabaseServer }
                     }
-                })
+                }
+                $spServiceApp = $spServiceApp | Add-Member ScriptMethod GetType { 
+                    return @{ FullName = $getTypeFullName } 
+                } -PassThru -Force
+                return $spServiceApp
             }
 
             It "returns values from the get method" {
@@ -90,16 +98,19 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
         }
 
         Context "When a service application exists and it should, but the app pool is not configured correctly" {
-            Mock Get-SPServiceApplication { 
-                return @(@{
-                    TypeName = "Business Data Connectivity Service Application"
+            Mock Get-SPServiceApplication {
+                $spServiceApp = [pscustomobject]@{
                     DisplayName = $testParams.Name
                     ApplicationPool = @{ Name = "Wrong App Pool Name" }
                     Database = @{
                         Name = $testParams.DatabaseName
                         Server = @{ Name = $testParams.DatabaseServer }
                     }
-                })
+                }
+                $spServiceApp = $spServiceApp | Add-Member ScriptMethod GetType { 
+                    return @{ FullName = $getTypeFullName } 
+                } -PassThru -Force
+                return $spServiceApp
             }
             Mock Get-SPServiceApplicationPool { return @{ Name = $testParams.ApplicationPool } }
             Mock Set-SPBusinessDataCatalogServiceApplication { }
@@ -122,16 +133,19 @@ Describe "SPBCSServiceApp - SharePoint Build $((Get-Item $SharePointCmdletModule
             Ensure = "Absent"
         }
         Context "When the service application exists but it shouldn't" {
-            Mock Get-SPServiceApplication { 
-                return @(@{
-                    TypeName = "Business Data Connectivity Service Application"
+            Mock Get-SPServiceApplication {
+                $spServiceApp = [pscustomobject]@{
                     DisplayName = $testParams.Name
                     ApplicationPool = @{ Name = $testParams.ApplicationPool }
                     Database = @{
                         Name = $testParams.DatabaseName
                         Server = @{ Name = $testParams.DatabaseServer }
                     }
-                })
+                }
+                $spServiceApp = $spServiceApp | Add-Member ScriptMethod GetType { 
+                    return @{ FullName = $getTypeFullName } 
+                } -PassThru -Force
+                return $spServiceApp
             }
             
             It "returns present from the Get method" {